vet-edit-report - extract the form part of submit-report-form into a general-report-form
